Improvements in data loading, added missing functions

+++ utils/train_utils.py
@@ -67,7 +67,7 @@ def get_dir_name(models_dir):
                     for d in os.listdir(models_dir)
                     if os.path.isdir(os.path.join(models_dir, d))
                     ]
-        ).astype(np.int)
+        ).astype(np.int32)
         if len(existing_dirs) > 0:
             dir_id = str(existing_dirs.max() + 1)
         else:
@@ -99,7 +99,7 @@ def get_dir_name_for_age_prediction(models_dir):
                     for d in os.listdir(models_dir)
                     if os.path.isdir(os.path.join(models_dir, d))
                     ]
-        ).astype(np.int)
+        ).astype(np.int16)
         if len(existing_dirs) > 0:
             dir_id = str(existing_dirs.max())
         else:
